Remove broken optional native dependencies
All native dependencies that fail to compile when installing have been
removed from the project.

One of these was a transitive dependency of `dnode`, which was an
unused direct dependency. It has been removed. It was not imported
anywhere, but it was referenced as a variable name in one unused
module. This module has also been removed.

Another was caused by the use of an outdated version of
`eth-keyring-controller`, which has been updated to a more up-to-date
in-range version without the broken transitive dependency.